added menu. only continue functional

+++ world/planet.h
@@ -3,9 +3,9 @@
 #define PLANETSIZE 64
 #include "point.h"
 struct Planet{
-	unsigned int radius;
-	int center_x;
-	int center_y;
+	float radius;
+	float center_x;
+	float center_y;
 	struct Point points[PLANETSIZE];
 };
 #endif
